Rename ConvertTask to FileTask
